feat: Conversation Sidebar code improvements
