+++ src/script/page/LeftSidebar/panels/Conversations/Conversations.tsx
@@ -22,13 +22,11 @@ import React, {useEffect, useState} from 'react';
 import {amplify} from 'amplify';
 import {container} from 'tsyringe';
 
-import {CircleCloseIcon, GroupIcon, Input, SearchIcon, StarIcon, InfoIcon} from '@wireapp/react-ui-kit';
+import {CircleCloseIcon, Input, SearchIcon} from '@wireapp/react-ui-kit';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {CallingCell} from 'Components/calling/CallingCell';
-import {Icon} from 'Components/Icon';
 import {Config} from 'src/script/Config';
-import {Conversation} from 'src/script/entity/Conversation';
 import {IntegrationRepository} from 'src/script/integration/IntegrationRepository';
 import {
   closeIconStyles,
@@ -48,7 +46,8 @@ import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {t} from 'Util/LocalizerUtil';
 
 import {ConversationsList} from './ConversationsList';
-import {ConversationTab} from './ConversationTab';
+import {ConversationTabs} from './ConversationTabs';
+import {getTabConversations} from './helpers';
 import {useFolderState} from './state';
 
 import {CallState} from '../../../../calling/CallState';
@@ -61,8 +60,6 @@ import {PreferenceNotificationRepository} from '../../../../notification/Prefere
 import {PropertiesRepository} from '../../../../properties/PropertiesRepository';
 import {PROPERTIES_TYPE} from '../../../../properties/PropertiesType';
 import {TeamState} from '../../../../team/TeamState';
-import {Shortcut} from '../../../../ui/Shortcut';
-import {ShortcutType} from '../../../../ui/ShortcutType';
 import {UserState} from '../../../../user/UserState';
 import {ListViewModel} from '../../../../view_model/ListViewModel';
 import {UserDetails} from '../../UserDetails';
@@ -78,7 +75,6 @@ type ConversationsProps = {
   selfUser: User;
   teamState?: TeamState;
   userState?: UserState;
-  listState: ListState;
   integrationRepository: IntegrationRepository;
   searchRepository: SearchRepository;
   teamRepository: TeamRepository;
@@ -110,39 +106,29 @@ const Conversations: React.FC<ConversationsProps> = ({
   callState = container.resolve(CallState),
   userState = container.resolve(UserState),
   selfUser,
-  listState,
 }) => {
   const [conversationsFilter, setConversationsFilter] = useState<string>('');
+  const {activeCalls} = useKoSubscribableChildren(callState, ['activeCalls']);
   const {classifiedDomains} = useKoSubscribableChildren(teamState, ['classifiedDomains']);
   const {connectRequests} = useKoSubscribableChildren(userState, ['connectRequests']);
-  const {activeConversation, unreadConversations} = useKoSubscribableChildren(conversationState, [
+  const {
+    activeConversation,
+    unreadConversations,
+    archivedConversations,
+    groupConversations,
+    directConversations,
+    visibleConversations: conversations,
+  } = useKoSubscribableChildren(conversationState, [
     'activeConversation',
     'archivedConversations',
+    'groupConversations',
+    'directConversations',
     'unreadConversations',
-  ]);
-
-  const {archivedConversations, visibleConversations: conversations} = useKoSubscribableChildren(conversationState, [
-    'archivedConversations',
     'visibleConversations',
   ]);
 
   const {conversationLabelRepository} = conversationRepository;
-
   const favoriteConversations = conversationLabelRepository.getFavorites(conversations);
-  const groupConversations = conversations.filter(conversation => conversation.isGroup());
-  const directConversations = conversations.filter(conversation => conversation.is1to1());
-
-  const totalUnreadConversations = unreadConversations.length;
-
-  const totalUnreadFavoriteConversations = favoriteConversations.filter(favoriteConversation =>
-    favoriteConversation.hasUnread(),
-  ).length;
-
-  const totalUnreadArchivedConversations = archivedConversations.filter(conversation =>
-    conversation.hasUnread(),
-  ).length;
-
-  const {activeCalls} = useKoSubscribableChildren(callState, ['activeCalls']);
 
   const initialTab = propertiesRepository.getPreference(PROPERTIES_TYPE.INTERFACE.VIEW_FOLDERS)
     ? SidebarTabs.FOLDER
@@ -150,19 +136,32 @@ const Conversations: React.FC<ConversationsProps> = ({
 
   const [currentTab, setCurrentTab] = useState<SidebarTabs>(initialTab);
 
-  const isRecentTab = currentTab === SidebarTabs.RECENT;
   const isFolderTab = currentTab === SidebarTabs.FOLDER;
-  const isFavoritesTab = currentTab === SidebarTabs.FAVORITES;
-  const isGroupsTab = currentTab === SidebarTabs.GROUPS;
-  const isDirectsTab = currentTab === SidebarTabs.DIRECTS;
-  const isArchivesTab = currentTab === SidebarTabs.ARCHIVES;
-  const isConnectTab = currentTab === SidebarTabs.CONNECT;
   const isPreferences = currentTab === SidebarTabs.PREFERENCES;
 
-  const showSearchInput = isRecentTab || isFolderTab || isFavoritesTab || isGroupsTab || isDirectsTab || isArchivesTab;
+  const showSearchInput = [
+    SidebarTabs.RECENT,
+    SidebarTabs.FOLDER,
+    SidebarTabs.FAVORITES,
+    SidebarTabs.GROUPS,
+    SidebarTabs.DIRECTS,
+    SidebarTabs.ARCHIVES,
+  ].includes(currentTab);
+
+  const {setCurrentView} = useAppMainState(state => state.responsiveView);
+  const {isOpen: isFolderOpen, openFolder} = useFolderState();
+  const {currentFocus, handleKeyDown, resetConversationFocus} = useConversationFocus(conversations);
+  const {conversations: currentTabConversations, searchInputPlaceholder} = getTabConversations({
+    currentTab,
+    conversations,
+    conversationsFilter,
+    archivedConversations,
+    groupConversations,
+    directConversations,
+    favoriteConversations,
+  });
 
   const hasNoConversations = conversations.length + connectRequests.length === 0;
-  const {isOpen: isFolderOpen, openFolder} = useFolderState();
 
   useEffect(() => {
     if (activeConversation && !conversationState.isVisible(activeConversation)) {
@@ -219,218 +218,6 @@ const Conversations: React.FC<ConversationsProps> = ({
     setCurrentTab(nextTab);
   }
 
-  const conversationTabs = [
-    {
-      type: SidebarTabs.RECENT,
-      title: t('conversationViewTooltip'),
-      dataUieName: 'go-recent-view',
-      Icon: <Icon.ConversationsOutline />,
-      unreadConversations: unreadConversations.length,
-    },
-    {
-      type: SidebarTabs.FAVORITES,
-      title: t('conversationLabelFavorites'),
-      dataUieName: 'go-favorites-view',
-      Icon: <StarIcon />,
-      unreadConversations: totalUnreadFavoriteConversations,
-    },
-    {
-      type: SidebarTabs.GROUPS,
-      title: t('conversationLabelGroups'),
-      dataUieName: 'go-groups-view',
-      Icon: <GroupIcon />,
-      unreadConversations: groupConversations.filter(conversation => conversation.hasUnread()).length,
-    },
-    {
-      type: SidebarTabs.DIRECTS,
-      title: t('conversationLabelDirects'),
-      dataUieName: 'go-directs-view',
-      Icon: <Icon.People />,
-      unreadConversations: directConversations.filter(conversation => conversation.hasUnread()).length,
-    },
-    {
-      type: SidebarTabs.FOLDER,
-      title: t('folderViewTooltip'),
-      dataUieName: 'go-folders-view',
-      Icon: <Icon.ConversationsFolder />,
-      unreadConversations: totalUnreadConversations,
-    },
-    {
-      hideTab: archivedConversations.length === 0,
-      type: SidebarTabs.ARCHIVES,
-      title: t('tooltipConversationsArchived', archivedConversations.length),
-      label: t('conversationFooterArchive'),
-      dataUieName: 'go-archive',
-      Icon: <Icon.Archive />,
-      unreadConversations: totalUnreadArchivedConversations,
-    },
-  ];
-
-  const sidebar = (
-    <nav className="conversations-sidebar">
-      <UserDetails
-        user={selfUser}
-        groupId={conversationState.selfMLSConversation()?.groupId}
-        isTeam={teamState.isTeam()}
-      />
-
-      <div
-        role="tablist"
-        aria-label={t('accessibility.headings.sidebar')}
-        aria-owns="tab-1 tab-2 tab-3 tab-4 tab-5 tab-6 tab-7"
-        className="conversations-sidebar-list"
-      >
-        <div className="conversations-sidebar-title">{t('videoCallOverlayConversations')}</div>
-
-        {conversationTabs.map((conversationTab, index) => {
-          if (conversationTab.hideTab) {
-            return null;
-          }
-
-          return (
-            <ConversationTab
-              {...conversationTab}
-              key={conversationTab.type}
-              conversationTabIndex={index + 1}
-              onChangeTab={changeTab}
-              isActive={conversationTab.type === currentTab}
-            />
-          );
-        })}
-
-        <div className="conversations-sidebar-title">{t('conversationFooterContacts')}</div>
-
-        <ConversationTab
-          title={t('searchConnect', Shortcut.getShortcutTooltip(ShortcutType.START))}
-          label={t('searchConnect')}
-          type={SidebarTabs.CONNECT}
-          Icon={<Icon.Plus />}
-          onChangeTab={changeTab}
-          conversationTabIndex={conversationTabs.length + 1}
-          dataUieName="go-people"
-          isActive={isConnectTab}
-        />
-      </div>
-
-      <div
-        role="tablist"
-        aria-label={t('accessibility.headings.sidebar.footer')}
-        aria-owns="tab-1 tab-2"
-        className="conversations-sidebar-list-footer"
-      >
-        <ConversationTab
-          title={t('preferencesHeadline', Shortcut.getShortcutTooltip(ShortcutType.START))}
-          label={t('preferencesHeadline')}
-          type={SidebarTabs.PREFERENCES}
-          Icon={<Icon.Settings />}
-          onChangeTab={tab => {
-            changeTab(tab);
-            onClickPreferences(ContentState.PREFERENCES_ACCOUNT);
-          }}
-          conversationTabIndex={1}
-          dataUieName="go-preferences"
-          isActive={isPreferences}
-        />
-
-        <a
-          rel="nofollow noopener noreferrer"
-          target="_blank"
-          href={Config.getConfig().URL.SUPPORT.INDEX}
-          id="tab-2"
-          type="button"
-          className="conversations-sidebar-btn"
-          title={t('preferencesAboutSupport', Shortcut.getShortcutTooltip(ShortcutType.START))}
-          data-uie-name="go-people"
-        >
-          <span className="conversations-sidebar-btn--text-wrapper">
-            <InfoIcon />
-            <span className="conversations-sidebar-btn--text">{t('preferencesAboutSupport')}</span>
-          </span>
-        </a>
-      </div>
-    </nav>
-  );
-
-  const callingView = (
-    <>
-      {activeCalls.map(call => {
-        const conversation = conversationState.findConversation(call.conversationId);
-        const callingViewModel = listViewModel.callingViewModel;
-        const callingRepository = callingViewModel.callingRepository;
-
-        return (
-          conversation && (
-            <div className="calling-cell" key={conversation.id}>
-              <CallingCell
-                classifiedDomains={classifiedDomains}
-                call={call}
-                callActions={callingViewModel.callActions}
-                callingRepository={callingRepository}
-                conversation={conversation}
-                isFullUi
-                hasAccessToCamera={callingViewModel.hasAccessToCamera()}
-                isSelfVerified={selfUser.is_verified()}
-                multitasking={callingViewModel.multitasking}
-              />
-            </div>
-          )
-        );
-      })}
-    </>
-  );
-
-  const {currentFocus, handleKeyDown, resetConversationFocus} = useConversationFocus(conversations);
-
-  function conversationSearchFilter(conversation: Conversation) {
-    return conversation.display_name().toLowerCase().includes(conversationsFilter.toLowerCase());
-  }
-
-  function getTabConversations() {
-    if ([SidebarTabs.FOLDER, SidebarTabs.RECENT].includes(currentTab)) {
-      return {
-        conversations: conversations.filter(conversationSearchFilter),
-        searchInputPlaceholder: t('searchConversations'),
-      };
-    }
-
-    if (currentTab === SidebarTabs.GROUPS) {
-      return {
-        conversations: groupConversations.filter(conversationSearchFilter),
-        searchInputPlaceholder: t('searchGroupConversations'),
-      };
-    }
-
-    if (currentTab === SidebarTabs.DIRECTS) {
-      return {
-        conversations: directConversations.filter(conversationSearchFilter),
-        searchInputPlaceholder: t('searchDirectConversations'),
-      };
-    }
-
-    if (currentTab === SidebarTabs.FAVORITES) {
-      return {
-        conversations: favoriteConversations.filter(conversationSearchFilter),
-        searchInputPlaceholder: t('searchFavoriteConversations'),
-      };
-    }
-
-    if (currentTab === SidebarTabs.ARCHIVES) {
-      return {
-        conversations: archivedConversations.filter(conversationSearchFilter),
-        searchInputPlaceholder: t('searchArchivedConversations'),
-      };
-    }
-
-    return {
-      conversations: [],
-      searchInputPlaceholder: '',
-    };
-  }
-
-  const {conversations: currentTabConversations, searchInputPlaceholder} = getTabConversations();
-
-  const {setCurrentView} = useAppMainState(state => state.responsiveView);
-
   const switchList = listViewModel.switchList;
 
   const onExitPreferences = () => {
@@ -451,102 +238,148 @@ const Conversations: React.FC<ConversationsProps> = ({
     }, ANIMATED_PAGE_TRANSITION_DURATION + 1);
   }
 
-  function getListWrapperItems() {
-    if (isPreferences) {
-      return (
-        <Preferences
-          onPreferenceItemClick={onClickPreferences}
-          teamRepository={teamRepository}
-          preferenceNotificationRepository={preferenceNotificationRepository}
-        />
-      );
-    }
+  const sidebar = (
+    <nav className="conversations-sidebar">
+      <UserDetails
+        user={selfUser}
+        groupId={conversationState.selfMLSConversation()?.groupId}
+        isTeam={teamState.isTeam()}
+      />
 
-    return hasNoConversations ? (
-      <>
-        {archivedConversations.length === 0 ? (
-          <div className="conversations-centered">
-            <div>
-              {t('conversationsWelcome', {
-                brandName: Config.getConfig().BRAND_NAME,
-              })}
-            </div>
-            <button className="button-reset-default text-underline" onClick={() => changeTab(SidebarTabs.CONNECT)}>
-              {t('conversationsNoConversations')}
-            </button>
+      <ConversationTabs
+        unreadConversations={unreadConversations}
+        favoriteConversations={favoriteConversations}
+        archivedConversations={archivedConversations}
+        groupConversations={groupConversations}
+        directConversations={directConversations}
+        onChangeTab={changeTab}
+        currentTab={currentTab}
+        onClickPreferences={() => onClickPreferences(ContentState.PREFERENCES_ACCOUNT)}
+      />
+    </nav>
+  );
+
+  const callingView = (
+    <>
+      {activeCalls.map(call => {
+        const conversation = conversationState.findConversation(call.conversationId);
+
+        if (!conversation) {
+          return null;
+        }
+
+        const callingViewModel = listViewModel.callingViewModel;
+        const {callingRepository} = callingViewModel;
+
+        return (
+          <div className="calling-cell" key={conversation.id}>
+            <CallingCell
+              classifiedDomains={classifiedDomains}
+              call={call}
+              callActions={callingViewModel.callActions}
+              callingRepository={callingRepository}
+              conversation={conversation}
+              isFullUi
+              hasAccessToCamera={callingViewModel.hasAccessToCamera()}
+              isSelfVerified={selfUser.is_verified()}
+              multitasking={callingViewModel.multitasking}
+            />
           </div>
-        ) : (
-          <div className="conversations-all-archived">{t('conversationsAllArchived')}</div>
-        )}
-      </>
-    ) : (
-      <>
-        {showSearchInput && (
-          <Input
-            className="label-1"
-            value={conversationsFilter}
-            onChange={event => {
-              setConversationsFilter(event.currentTarget.value);
-            }}
-            startContent={<SearchIcon width={14} height={14} css={searchIconStyles} />}
-            endContent={
-              conversationsFilter && (
-                <CircleCloseIcon
-                  className="cursor-pointer"
-                  onClick={() => setConversationsFilter('')}
-                  css={closeIconStyles}
-                />
-              )
-            }
-            inputCSS={searchInputStyles}
-            wrapperCSS={searchInputWrapperStyles}
-            placeholder={searchInputPlaceholder}
-          />
-        )}
-        {showSearchInput && currentTabConversations.length === 0 && (
-          <div className="conversations-centered">
-            <div>{t('searchConversationsNoResult')}</div>
-            <button className="button-reset-default text-underline" onClick={() => changeTab(SidebarTabs.CONNECT)}>
-              {t('searchConversationsNoResultConnectSuggestion')}
-            </button>
-          </div>
-        )}
-
-        {isConnectTab && (
-          <StartUI
-            conversationRepository={conversationRepository}
-            searchRepository={searchRepository}
-            teamRepository={teamRepository}
-            integrationRepository={integrationRepository}
-            mainViewModel={listViewModel.mainViewModel}
-            userRepository={userRepository}
-            isFederated={listViewModel.isFederated}
-            selfUser={selfUser}
-          />
-        )}
-
-        {showSearchInput && (
-          <ConversationsList
-            callState={callState}
-            currentTab={currentTab}
-            currentFocus={currentFocus}
-            listViewModel={listViewModel}
-            connectRequests={connectRequests}
-            handleArrowKeyDown={handleKeyDown}
-            conversationState={conversationState}
-            conversations={currentTabConversations}
-            conversationRepository={conversationRepository}
-            resetConversationFocus={resetConversationFocus}
-          />
-        )}
-      </>
-    );
-  }
+        );
+      })}
+    </>
+  );
 
   return (
     <div className="conversations-wrapper">
       <ListWrapper id="conversations" headerElement={<></>} hasHeader={false} sidebar={sidebar} before={callingView}>
-        {getListWrapperItems()}
+        {isPreferences && (
+          <Preferences
+            onPreferenceItemClick={onClickPreferences}
+            teamRepository={teamRepository}
+            preferenceNotificationRepository={preferenceNotificationRepository}
+          />
+        )}
+
+        {isPreferences ? null : hasNoConversations ? (
+          <>
+            {archivedConversations.length === 0 ? (
+              <div className="conversations-centered">
+                <div>
+                  {t('conversationsWelcome', {
+                    brandName: Config.getConfig().BRAND_NAME,
+                  })}
+                </div>
+                <button className="button-reset-default text-underline" onClick={() => changeTab(SidebarTabs.CONNECT)}>
+                  {t('conversationsNoConversations')}
+                </button>
+              </div>
+            ) : (
+              <div className="conversations-all-archived">{t('conversationsAllArchived')}</div>
+            )}
+          </>
+        ) : (
+          <>
+            {showSearchInput && (
+              <Input
+                className="label-1"
+                value={conversationsFilter}
+                onChange={event => {
+                  setConversationsFilter(event.currentTarget.value);
+                }}
+                startContent={<SearchIcon width={14} height={14} css={searchIconStyles} />}
+                endContent={
+                  conversationsFilter && (
+                    <CircleCloseIcon
+                      className="cursor-pointer"
+                      onClick={() => setConversationsFilter('')}
+                      css={closeIconStyles}
+                    />
+                  )
+                }
+                inputCSS={searchInputStyles}
+                wrapperCSS={searchInputWrapperStyles}
+                placeholder={searchInputPlaceholder}
+              />
+            )}
+            {showSearchInput && currentTabConversations.length === 0 && (
+              <div className="conversations-centered">
+                <div>{t('searchConversationsNoResult')}</div>
+                <button className="button-reset-default text-underline" onClick={() => changeTab(SidebarTabs.CONNECT)}>
+                  {t('searchConversationsNoResultConnectSuggestion')}
+                </button>
+              </div>
+            )}
+
+            {currentTab === SidebarTabs.CONNECT && (
+              <StartUI
+                conversationRepository={conversationRepository}
+                searchRepository={searchRepository}
+                teamRepository={teamRepository}
+                integrationRepository={integrationRepository}
+                mainViewModel={listViewModel.mainViewModel}
+                userRepository={userRepository}
+                isFederated={listViewModel.isFederated}
+                selfUser={selfUser}
+              />
+            )}
+
+            {showSearchInput && (
+              <ConversationsList
+                callState={callState}
+                currentTab={currentTab}
+                currentFocus={currentFocus}
+                listViewModel={listViewModel}
+                connectRequests={connectRequests}
+                handleArrowKeyDown={handleKeyDown}
+                conversationState={conversationState}
+                conversations={currentTabConversations}
+                conversationRepository={conversationRepository}
+                resetConversationFocus={resetConversationFocus}
+              />
+            )}
+          </>
+        )}
       </ListWrapper>
     </div>
   );
